add test and refactorG
