+++ src/users/get-users.ts
@@ -1,4 +1,4 @@
-import { get } from '../common/request';
+import { getPages } from '../common/get-pages';
 import { IAuth0AuthorizationApiUser } from '../interfaces';
 
 export interface Response {
@@ -25,26 +25,6 @@ export function getUsers(authorizationExtensionUrl: string, accessToken: string)
       ...getUsersDefaultOptions,
       ...options
     };
-
-    const getPaged: any = function(page: number) {
-      return get({
-        accessToken,
-        url: `${authorizationExtensionUrl}/users`,
-        queryParams: {
-          page: `${options.page}`,
-          per_page: `${options.perPage}`,
-        }
-      }).then((result: any) => {
-        const total_pages = Math.ceil(result.total / options.perPage);
-        if (total_pages > page + 1) {
-          return getPaged(page + 1).then((p: any) => {
-            result.users = result.users.concat(p.users);
-            return result;
-          })
-        }
-        return result;
-      });
-    };
-    return getPaged(0);
+    return getPages(`${authorizationExtensionUrl}/users`, accessToken, options);
   }
 }
